fix: убирает связь данных с редусером - начальное состояние приложения передаётся в стор в момент создания
