+++ src/components/movie-nav-genre/movie-nav-genre.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import MovieNavGenre from './movie-nav-genre.jsx';
+import {films} from '../data-for-test.js';
 
 const MovieGenreList = [
   `All genres`,
@@ -19,6 +20,7 @@ describe(`MovieNavGenre`, () => {
   it(`Render MovieNavGenre`, () => {
     const tree = renderer.create(
         <MovieNavGenre
+          films={films}
           genres={MovieGenreList}
           currentGenre={MovieGenreList[0]}
           onGenreClick={() => {}}
