+++ src/index.js
@@ -6,9 +6,23 @@ import {Provider} from 'react-redux';
 import {reducer} from './store/reducer.js';
 import App from './components/app/app.jsx';
 
+import {ALL_GENRES} from './const.js';
+import {getAvailableGenres} from './utils.js';
+import films from './mocks/films.js';
+import moviePoster from './mocks/movie-poster.js';
+
 const ENTRY_POINT = document.querySelector(`#root`);
+
+const initialState = {
+  films,
+  moviePoster,
+  availableGenres: getAvailableGenres(films),
+  currentGenre: ALL_GENRES,
+  filmsByGenre: films,
+};
+
 const store = createStore(
-    reducer,
+    reducer, initialState,
     window.__REDUX_DEVTOOLS_EXTENSION__ ? window.__REDUX_DEVTOOLS_EXTENSION__() : (f) => f);
 
 ReactDom.render(
