+++ src/store/reducer.test.js
@@ -1,6 +1,4 @@
-import {getAvailableGenres} from '../utils.js';
-import films from '../mocks/films.js';
-import moviePoster from '../mocks/movie-poster.js';
+import {films} from '../components/data-for-test.js';
 import {ActionType, reducer} from './reducer.js';
 
 const getFilmsByGenre = (selectedGenre) => {
@@ -8,16 +6,6 @@ const getFilmsByGenre = (selectedGenre) => {
 };
 
 describe(`Reducer`, () => {
-  it(`Return initial state`, () => {
-    expect(reducer(void 0, {})).toEqual({
-      films,
-      moviePoster,
-      availableGenres: getAvailableGenres(films),
-      currentGenre: `All genres`,
-      filmsByGenre: films,
-    });
-  });
-
   it(`Return genre after choise`, () => {
     expect(reducer({
       currentGenre: `All genres`,
